% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dipper.R
\name{plot_dipper}
\alias{plot_dipper}
\title{Create a forest plot from a DiPPER fit}
\usage{
plot_dipper(fit, prob = 0.95, show_taxa = "significant")
}
\arguments{
\item{fit}{A dipper_fit object returned by run_dipper.}

\item{prob}{Numeric. Probability mass for the credible intervals.
Default is 0.95 (95\% interval).}

\item{show_taxa}{Selection criteria for taxa: "significant" (default),
"all", or an integer k (top k taxa ranked by pseudo_q).}
}
\value{
A ggplot object.
}
\description{
Create a forest plot from a DiPPER fit
}
