% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_dipper.R
\name{summary_dipper}
\alias{summary_dipper}
\title{Summarize the results from a DiPPER model fit}
\usage{
summary_dipper(fit, prob = 0.95, scale = c("log_odds", "odds_ratio"))
}
\arguments{
\item{fit}{A dipper_fit object returned by run_dipper.}

\item{prob}{Probability mass for the credible interval (default 0.95).}

\item{scale}{Character string, either "log_odds" or "odds_ratio".}
}
\value{
A data.frame containing the summarized results for each taxon.
}
\description{
Summarize the results from a DiPPER model fit
}
