% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tse_hintikka}
\alias{tse_hintikka}
\title{Gut microbiome dataset from Hintikka et al. (2021)}
\format{
A \code{TreeSummarizedExperiment} object.
\describe{
\item{assay}{Contains the \code{counts} matrix.}
\item{colData}{Contains the sample metadata, including \code{Fat}
(diet group, factor with levels Low and High) and \code{XOS}
(xylo-oligosaccharide supplementation).}
}
}
\source{
Derived from the \code{HintikkaXOData} dataset in the
\code{mia} package. See \code{inst/scripts/tse_hintikka.R} for how the
subset was derived. Original study:
\url{https://www.mdpi.com/1660-4601/18/8/4049}
}
\usage{
data(tse_hintikka)
}
\description{
A \code{TreeSummarizedExperiment} object containing a subset of gut
microbiome data from an experiment studying the effects of a high-fat
diet and xylo-oligosaccharide (XOS) supplementation on rats (N = 20 + 20).
The abundance data are sequencing counts and they have been agglomerated
to the genus level.
}
\keyword{datasets}
