% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiPPER.R
\name{DiPPER}
\alias{DiPPER}
\title{Run the DiPPER model}
\usage{
DiPPER(
  tse = NULL,
  meta = NULL,
  abund_matrix = NULL,
  formula,
  var_of_interest = NULL,
  read_depth = TRUE,
  symmetric = F,
  assay_type = c("counts", "relabundance", "pa"),
  assay_name = NULL,
  tax_rank = NULL,
  pa_threshold = 0,
  min_present = 5,
  min_absent = min_present,
  iter_sampling = 1000,
  cores = 4,
  seed = 1,
  ...
)
}
\arguments{
\item{tse}{A TreeSummarizedExperiment object.}

\item{meta}{A data.frame containing metadata (required if tse is NULL).}

\item{abund_matrix}{A matrix containing counts, relative abundances, or
presence/absence data (required if tse is NULL).}

\item{formula}{Model formula.}

\item{var_of_interest}{The variable of interest from the formula.
If NULL, the first term of the formula is automatically used.}

\item{read_depth}{Logical or character. If TRUE, calculates log10 read depth
from counts. If character, uses the specified metadata column.}

\item{symmetric}{Logical. If TRUE, a symmetric Laplace prior (nu = 0.5)
is used for the differential prevelance parameters of interest. If FALSE
(default), the prior is allowed to be asymmetric.}

\item{assay_type}{Character. Type of the assay: "counts" (default),
"relabundance", or "pa" (presence/absence).}

\item{assay_name}{Name of the assay in the TSE object to use. If NULL
(default), it is automatically determined based on assay_type. Ignored if
tse is NULL.}

\item{tax_rank}{A character string specifying the taxonomic rank to
agglomerate by (e.g., "genus"). If NULL, no agglomeration is done.
Ignored if tse is NULL.}

\item{pa_threshold}{Numeric or function. Threshold for presence. If strictly
between 0 and 1, relative abundances are used. If >= 1 or exactly 0,
counts are used. If a function (e.g., mean), it is applied per taxon.
Default is 0. Ignored if assay_type is "pa".}

\item{min_present}{Minimum number or proportion of samples where a taxon
must be present.}

\item{min_absent}{Minimum number or proportion of samples where a taxon
must be absent. Default is min_present.}

\item{iter_sampling}{Number of post-warmup MCMC iterations per chain.}

\item{cores}{Number of CPU cores to use.}

\item{seed}{Random seed for reproducibility.}

\item{...}{Additional arguments passed directly to \code{run_dipper}.}
}
\value{
The output of the \code{run_dipper} function.
}
\description{
This is a wrapper function that sequentially runs \code{prep_dipper_data}
and \code{run_dipper}.
}
