% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dipper.R
\name{dipper}
\alias{dipper}
\title{Run DiPPER}
\usage{
dipper(
  tse = NULL,
  formula,
  assay = NULL,
  meta = NULL,
  assay.type = NULL,
  data.type = c("counts", "relabundance", "pa"),
  var.of.interest = NULL,
  read.depth = TRUE,
  symmetric = FALSE,
  threshold = 0,
  min.present = 5,
  min.absent = min.present,
  iter.sampling = 1000,
  iter.warmup = iter.sampling,
  chains = 4,
  cores = 4,
  adapt.delta = 0.8,
  max.treedepth = 10,
  run.diagnostics = TRUE,
  diagnostics.level = c("basic", "full"),
  seed = 1,
  print.progress = 200,
  prior.alpha.sd = 4,
  prior.tau.sd = 1,
  prior.nu.sd = 0.05,
  prior.cov.sd = 1,
  prior.reads.mean = 2,
  prior.reads.sd = 2,
  prior.sigma.subj = 1,
  ...
)
}
\arguments{
\item{tse}{A TreeSummarizedExperiment object.}

\item{formula}{Model formula.}

\item{assay}{A matrix containing counts, relative abundances, or
presence/absence data (required if tse is NULL). Rows should be
taxa/features, columns should be samples.}

\item{meta}{A data.frame containing metadata (required if tse is NULL).}

\item{assay.type}{Character. Name of the assay in the TSE object to use.
Required if tse is provided.}

\item{data.type}{Character. Type of the data: "counts" (default),
"relabundance", or "pa" (presence/absence).}

\item{var.of.interest}{The variable of interest from the formula.
If NULL, the first term of the formula is automatically used.}

\item{read.depth}{Logical or character. If TRUE, calculates log10 read depth
from counts. If character, uses the specified metadata column.}

\item{symmetric}{Logical. If TRUE, a symmetric Laplace prior (nu = 0.5)
is used for the differential prevalence parameters of interest. If FALSE
(default), the prior is allowed to be asymmetric.}

\item{threshold}{Numeric or function. Threshold for presence.
If between 0 and 1, the threshold is based on relative abundance (e.g.,
0.05 means that relative abundances > 0.05 are considered present).
If whole number, the threshold is based on counts (e.g. 5 means that
count > 5 is considered present).
If a function (e.g. median), it is applied for the given assay for each
feature. Default is 0. Ignored if data.type is "pa".}

\item{min.present}{For taxon/feature filtering. Minimum number or proportion
of samples where a taxon must be present. Default is 5.}

\item{min.absent}{For taxon/feature filtering. Minimum number or proportion
of samples where a taxon must be absent. Default is min.present.}

\item{iter.sampling}{Number of post-warmup MCMC iterations per chain.
Default is 1000.}

\item{iter.warmup}{Number of warmup MCMC iterations per chain.
Defaults to iter.sampling.}

\item{chains}{Number of MCMC chains. Default is 4.}

\item{cores}{Number of CPU cores to use. Default is 4.}

\item{adapt.delta}{Target average acceptance probability. Default is 0.8.}

\item{max.treedepth}{Maximum depth of the trees. Default is 10.}

\item{run.diagnostics}{Logical. Whether to run convergence diagnostics
after sampling. Default is TRUE.}

\item{diagnostics.level}{Character. Level of MCMC diagnostics to run.
"basic" (default) checks main parameters (alpha, beta, covariates,
hyperparameters).
"full" checks all parameters, including latent and auxiliary variables.}

\item{seed}{Random seed for reproducibility. Default is 1.}

\item{print.progress}{How often to print MCMC progress. Set to 0 or
FALSE to disable. Default is 200.}

\item{prior.alpha.sd}{Prior standard deviation for alpha (fixed intercept
for centered variables). Default is 4.0.}

\item{prior.tau.sd}{Prior standard deviation for tau (global scale of the
Symmetric/Asymmetric Laplace prior). Default is 1.0.}

\item{prior.nu.sd}{Prior standard deviation for nu (asymmetry parameter).
Ignored if symmetric = TRUE. Default is 0.05.}

\item{prior.cov.sd}{Default prior standard deviation for covariates.
Default is 1.0.}

\item{prior.reads.mean}{Prior mean for the read depth covariate.
Default is 2.0.}

\item{prior.reads.sd}{Prior standard deviation for the read depth covariate.
Default is 2.0.}

\item{prior.sigma.subj}{Prior standard deviation for the half-normal prior
distribution of the taxon-specific random intercept standard deviations.
Default is 1.0.}

\item{...}{Additional arguments passed directly to \code{run_dipper}
(e.g., cmdstanr specific arguments).}
}
\value{
A list object of class "dipper_fit".
}
\description{
This is a wrapper function that sequentially runs \code{prep_dipper_data}
and \code{run_dipper}.
}
