% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fit_example}
\alias{fit_example}
\title{Example model fit object for DiPPER}
\format{
A list of class \code{dipper_fit} containing:
\describe{
\item{stanfit}{The \code{CmdStanMCMC} object returned by CmdStanR.}
\item{dipper_data}{A list containing the prepared data passed to Stan.}
\item{symmetric}{Logical indicating if a symmetric Laplace prior for
differential prevalence parameters was used.}
}
}
\source{
Generated using the \code{tse_hintikka} dataset with
\code{niter = 400} and \code{chains = 1}.
}
\usage{
data("fit_example")
}
\description{
A pre-calculated \code{dipper_fit} object to be used in examples and
vignettes to avoid long MCMC sampling times during package checks.
}
\keyword{datasets}
