% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dipper_fit.R
\name{plot.dipper_fit}
\alias{plot.dipper_fit}
\title{Plot method for DiPPER fit objects}
\usage{
\method{plot}{dipper_fit}(x, prob = 0.95, show.taxa = 10, original.scale = TRUE, ...)
}
\arguments{
\item{x}{A dipper_fit object returned by run_dipper.}

\item{prob}{Numeric. Probability mass for the credible intervals.
Default is 0.95 (95\% interval).}

\item{show.taxa}{Selection criteria for taxa: "significant" (credible
intervals not overlapping OR = 1),
"all", or an integer k (top k taxa ranked by pseudo_q). Default is 10.}

\item{original.scale}{Logical. If TRUE (default), continuous covariates
are back-transformed to their original scale. If FALSE, results are
presented per one standard deviation increase.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
A ggplot object.
}
\description{
Plot method for DiPPER fit objects
}
