% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_dipper_data.R
\name{prep_dipper_data}
\alias{prep_dipper_data}
\title{Prepare data for DiPPER}
\usage{
prep_dipper_data(
  tse = NULL,
  formula,
  assay = NULL,
  meta = NULL,
  assay.type = NULL,
  data.type = c("counts", "relabundance", "pa"),
  var.of.interest = NULL,
  read.depth = TRUE,
  threshold = 0,
  min.present = 5,
  min.absent = min.present
)
}
\arguments{
\item{tse}{A TreeSummarizedExperiment object.}

\item{formula}{Model formula.}

\item{assay}{A matrix containing counts, relative abundances, or
presence/absence data (required if tse is NULL). Rows should be
taxa/features, columns should be samples.}

\item{meta}{A data.frame containing metadata (required if tse is NULL).}

\item{assay.type}{Character. Name of the assay in the TSE object to use.
Required if tse is provided.}

\item{data.type}{Character. Type of the data: "counts" (default),
"relabundance", or "pa" (presence/absence).}

\item{var.of.interest}{The variable of interest from the formula.
If NULL, the first term of the formula is automatically used.}

\item{read.depth}{Logical or character. If TRUE, calculates log10 read depth
from counts. If character, uses the specified metadata column.}

\item{threshold}{Numeric or function. Threshold for presence.
If between 0 and 1, the threshold is based on relative abundance (e.g.,
0.05 means that relative abundances > 0.05 are considered present).
If whole number, the threshold is based on counts (e.g. 5 means that
count > 5 is considered present).
If a function (e.g. median), it is applied for the given assay for each
feature. Default is 0. Ignored if data.type is "pa".}

\item{min.present}{For taxon/feature filtering. Minimum number or proportion
of samples where a taxon must be present. Default is 5.}

\item{min.absent}{For taxon/feature filtering. Minimum number or proportion
of samples where a taxon must be absent. Default is min.present.}
}
\description{
Prepare data for DiPPER
}
\keyword{internal}
