% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dipper_fit.R
\name{print.dipper_fit}
\alias{print.dipper_fit}
\title{Print a summary of a DiPPER model fit}
\usage{
\method{print}{dipper_fit}(x, ...)
}
\arguments{
\item{x}{A \code{dipper_fit} object.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
The object \code{x} invisibly.
}
\description{
Print a summary of a DiPPER model fit
}
\examples{
# Load pre-run model fit for the example dataset (tse_hintikka)
data("fit_example")

print(fit_example)
}
