% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_dipper.R
\name{run_dipper}
\alias{run_dipper}
\title{Run the DiPPER model using cmdstanr}
\usage{
run_dipper(
  prep_data,
  iter_sampling = 1000,
  iter_warmup = iter_sampling,
  run_diagnostics = TRUE,
  chains = 4,
  cores = 4,
  seed = 1,
  adapt_delta = 0.8,
  max_treedepth = 10,
  prior_nu_sd = 0.05,
  prior_tau_sd = 1,
  prior_alpha_sd = 4,
  prior_cov_sd = 1,
  prior_reads_mean = 2,
  prior_reads_sd = 2,
  print_progress = 200
)
}
\arguments{
\item{prep_data}{The list returned by \code{prep_dipper_data}.}

\item{iter_sampling}{Number of post-warmup MCMC iterations per chain.}

\item{iter_warmup}{Number of warmup MCMC iterations per chain. Defaults
to iter_sampling.}

\item{run_diagnostics}{Logical. Whether to run convergence diagnostics.
Default is TRUE.}

\item{chains}{Number of MCMC chains.}

\item{cores}{Number of CPU cores to use.}

\item{seed}{Random seed for reproducibility.}

\item{adapt_delta}{Target average acceptance probability (default 0.8).}

\item{max_treedepth}{Maximum depth of the trees (default 10).}

\item{prior_nu_sd}{Prior standard deviation for nu.}

\item{prior_tau_sd}{Prior standard deviation for tau.}

\item{prior_alpha_sd}{Prior standard deviation for alpha (intercept).}

\item{prior_cov_sd}{Default prior standard dev for covariates.}

\item{prior_reads_mean}{Prior mean for the read depth covariate.}

\item{prior_reads_sd}{Prior standard dev for the read depth covariate.}

\item{print_progress}{How often to print MCMC progress. Set to 0 or
FALSE to disable. Default is 200.}
}
\description{
Run the DiPPER model using cmdstanr
}
