% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tse_hintikka}
\alias{tse_hintikka}
\title{Rat microbiome data: low/high-fat diet and XOS supplementation}
\format{
A \code{TreeSummarizedExperiment} object with 40 samples (rats).
\describe{
\item{colData}{Contains experimental metadata. The included variables are
the diet group (\code{Fat}: High/Low) and supplementation (\code{XOS}:
No / Yes).}
\item{assays}{Contains the \code{counts} assay with microbial abundances.}
}
}
\source{
Hintikka L. et al. (2021). Xylo-Oligosaccharides in Prevention of
Hepatic Steatosis and Adipose Tissue Inflammation: Associating Taxonomic and
Metabolomic Patterns in Fecal Microbiomes with Biclustering.
\emph{International Journal of Environmental Research and Public Health},
18(8):4049; https://doi.org/10.3390/ijerph18084049.
}
\usage{
tse_hintikka
}
\description{
A \code{TreeSummarizedExperiment} object containing microbiome data from a
rat experiment studying the effects of a high-fat diet and prebiotic
xylo-oligosaccharide (XOS) supplementation.
}
\details{
The dataset has been agglomerated to the Genus level using
\code{mia::agglomerateByRank()}.
}
\keyword{datasets}
