% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VatanenT_2016_subset}
\alias{VatanenT_2016_subset}
\title{Longitudinal gut microbiome dataset from Vatanen et al. (2016)}
\format{
A \code{TreeSummarizedExperiment} object.
\describe{
\item{assay}{Contains the \code{relative_abundance} matrix (proportions).}
\item{colData}{Contains the sample metadata: \code{subject_id} (factor),
\code{antibiotics} (factor: no, yes), \code{age_point} (factor:
1.5 months, 7 months), and \code{gender} (factor: male, female).}
\item{rowTree}{Contains the phylogenetic tree.}
}
}
\source{
Data obtained via the \code{curatedMetagenomicData} R package.
Original study: \url{https://pmc.ncbi.nlm.nih.gov/articles/PMC4950857/}
}
\usage{
data(VatanenT_2016_subset)
}
\description{
A \code{TreeSummarizedExperiment} object containing a longitudinal subset of
infant gut microbiome data from Vatanen et al. (2016). The data consists of
79 infants with measurements at 1.5 and/or 7 months of age. The abundance
data are relative abundances (proportions) and they have been agglomerated
to the genus level.
}
\keyword{datasets}
